made suer only super users can navigate to add/edit/delete product page

+++ shop/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 from .models import Product
 from .forms import ProductForm
@@ -21,8 +22,13 @@ def product_page(request, product_id):
     }
     return render(request, 'shop/product.html', context)
 
+@login_required
 def add_product(request):
     """ A view for super users to add products """
+    if not request.user.is_superuser:
+        messages.error(request, 'Unfortunately only the super user can add/update/delete products')
+        return redirect(reverse('home'))
+        
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -39,8 +45,13 @@ def add_product(request):
         }
     return render(request, template, context)
 
-def edit_product(request, product_id): 
+@login_required
+def edit_product(request, product_id):
     """ A view for super users to edit products """
+    if not request.user.is_superuser:
+        messages.error(request, 'Unfortunately only the super user can add/update/delete products')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product,  pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -61,8 +72,13 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+@login_required
 def delete_product(request, product_id):
     """ A view for super users to delete products """
+    if not request.user.is_superuser:
+        messages.error(request, 'Unfortunately only the super user can add/update/delete products')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product,  pk=product_id)
     product.delete()
     messages.success(request, 'The product has been deleted successfully')
